+++ golang/kurtosis_module_docker_api/kurtosis_module_docker_api.go
@@ -18,8 +18,8 @@
 package kurtosis_module_docker_api
 
 const (
-	// IP:port of the Kurtosis API container
-	ApiContainerSocketEnvVar = "API_CONTAINER_SOCKET"
+	// IP:port of the Kurtosis engine server
+	EngineSocketEnvVar = "ENGINE_SOCKET"
 
 	// Arbitrary serialized data that the module can consume at startup to modify its behaviour
 	// Analogous to the "constructor"